refactor: extract CLI commands from monolithic main.go
Split cmd/devlog/main.go into per-command files and shared helpers so
each command is easier to navigate and test. main.go is now dispatch
and help only.

Fixes #27
